make language clearer
As it stands, there might be a lack of clarity about an "entry". We want to be contributors to be clear when they are asking for a *list* (rather than an identifier) that isn't currently present on org-id. Also, some of the requests we get via Github are for a specific organization identifier - sometimes they are on a current entry and sometimes not.